refactor(checker): split assignment widening display helpers
