+++ crates/tsz-checker/src/error_reporter/core/diagnostic_source.rs
@@ -2,6 +2,7 @@
 
 mod assignment_formatting;
 mod assignment_source_preservation;
+mod assignment_widening;
 mod compound_assignment_context;
 mod computed_index_source_display;
 mod contextual_index_display;
